Created macro to plot xADC test data
Pulls data from TTree and saves it in TH1D histograms for
each PDAC value, then grabs means and standard deviations
and stores in a TGraphErrors plot. At first I fitted it with
a linear fit, but now it's a quadratic fit because that was much
closer and fit more data points.  The quadratic correction is
~4 orders of magnitude smaller than the linear term.